+++ src/common.ts
@@ -66,7 +66,9 @@ export function parseArgs<T>(argv: string[], defaultData: IArguments<T> | (() =>
 
       if (columns[1] === 'true' || columns[1] === 'false') value = columns[1] === 'true';
 
-      // if (typeof defaultData[name] !== typeof value) throw new Error('Unexpected type \"' + typeof value + '\" for argument \"' + name + '\". The argument should be of type \"' + typeof defaultData[name] + '\".');
+      const defaultArgument = defaultData.arguments[name];
+
+      if (defaultArgument !== null && defaultArgument !== undefined && typeof defaultData.arguments[name] !== typeof value) throw new Error('Unexpected type \"' + typeof value + '\" for argument \"' + name + '\". The argument should be of type \"' + typeof defaultData[name] + '\".');
     } else {
       if (!commandFound) {
         if (!validArguments.includes(name)) throw new Error('The specified command \"' + name + '\" is invalid.');