+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 1.18.0
 ### Changes
 * `launch init` now updates the **package.json** file
+* `launch init` now supports templates
+* Added integration tests
 
 ## 1.17.3
 ### Changes